# apps/frontend/openapi.yaml
openapi: 3.1.0
info:
  title: ALL-Flow API
  version: 1.0.0
  description: |
    ALL-Flow 사내 협업 시스템 백엔드 API 스펙.
    프론트엔드 `lib/types.ts` 와 양방향으로 동기화됩니다.

    인증: Bearer JWT (next-auth v5 발급) — 모든 보호된 엔드포인트에 필요.
servers:
  - url: https://api.allflow.omelet.ai/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      tags: [tasks]
      summary: 태스크 목록
      parameters:
        - { name: projectId, in: query, schema: { type: string } }
        - { name: assigneeId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/StatusKey' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      tags: [tasks]
      summary: 태스크 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [tasks]
      summary: 태스크 부분 수정
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
    delete:
      tags: [tasks]
      summary: 태스크 삭제
      responses:
        '204': { description: Deleted }
...
  /reports/weekly:
    post:
      tags: [reports]
      summary: AI 주간 보고 자동 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [periodStart, periodEnd, scopeIds]
              properties:
                periodStart: { type: string, format: date }
                periodEnd: { type: string, format: date }
                scopeIds: { type: array, items: { type: string } }
                tone: { type: string, enum: [exec, team, casual] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
  /reports/monthly:
    post:
      tags: [reports]
      summary: AI 월간 보고 자동 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [year, month]
              properties:
                year: { type: integer }
                month: { type: integer, minimum: 1, maximum: 12 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
  /reports/{id}/send:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [reports]
      summary: 리포트 이메일 발송 큐 적재
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [recipients]
              properties:
                recipients:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: { type: string, format: email }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [queued, recipients]
                properties:
                  queued: { type: integer }
                  recipients:
                    type: array
                    items: { type: string, format: email }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StatusKey:
      type: string
      enum: [todo, doing, review, done, blocked]
...
    Task:
      type: object
      required: [id, title, status, proj, assignee, due, priority, tags]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/StatusKey' }
        proj: { type: string }
        assignee: { type: string }
        due: { type: string }
        priority:
          type: string
          enum: [high, med, low]
        tags:
          type: array
          items: { type: string }
    TaskCreate:
      type: object
      required: [title, proj, assignee]
      properties:
        title: { type: string }
        proj: { type: string }
        assignee: { type: string }
        due: { type: string, format: date }
        priority:
          type: string
          enum: [high, med, low]
    TaskPatch:
      type: object
      properties:
        title: { type: string }
        status: { $ref: '#/components/schemas/StatusKey' }
        assignee: { type: string }
        due: { type: string, format: date }
...
    Report:
      type: object
      required: [id, kind, periodStart, periodEnd, generatedAt, sections]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [weekly, monthly]
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        generatedAt: { type: string, format: date-time }
        author: { type: string }
        tldr: { type: string }
        kpis:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              value: { type: string }
              delta: { type: string }
              dir: { type: string, enum: [up, down, flat] }
        sections:
          type: array
          items:
            type: object
            properties:
              heading: { type: string }
              body: { type: string }
              citations:
                type: array
                items:
                  type: object
                  properties:
                    kind: { type: string }
                    id: { type: string }
                    label: { type: string }
...